Make pprof server use more intentional
Move the pprof server code into its own package, `debugserver`, which
makes it easier to test and contain.

Also manually assign pprof endpoints rather than use the blank import so
that they the endpoints are not added to the `DefaultServeMux`. Adding
to the `DefaultServeMux` semeed somewhat risky as we may accidentally
use that in other code.