Refactor cmd/system-metrics-agent
Rename system-metrics-agent to agentd. The module is already called
system-metrics, so we don't need to specify that again for the name of
the application. The `d` at the end indicates that this a daemon
application, to be run in the background.

Move the remaining code in the application dir to the main function and
metricserver package. The app dir inside the application dir wasn't
serving much of a purpose and could be split out for clarity and unit
testing.