change default metrics port
This port has been 53035 in all installations we're aware of since it
was integrated into cf-deployment. The ops files which include the
metrics agent set it to 53035.

We are changing this to free up port 9100 for use in the otel-collector
integration. We need a port which will be reserved on tcp router, and
those are few and far between, in fact we believe this is the only one.

It is possible someone is relying on this default, but we think that is
unlikely and the change is worth the risk. The alternative is a breaking
change in tcp router which will require all platform teams  to manually verify
that the port being claimed for otel-collector is unused and migrate
anything off of that port, which could be a breaking change for
application operators.

Signed-off-by: Rebecca Roberts <robertsre@vmware.com>